Waveform coordinates to wave file convertor
This patch was created as part of my Youtube video about converting an SVG vector file to a waveform for oscilloscope viewing.

It is a one-step solution to converting X/Y coordinate pairs to a waveform that can be rotated in an orthogonal, saved to disc and I added a feature that "writes" the waveform on screen, than erases it. See notes in the patch for details.

The conversion Pure Data code is from Derek Holzer's Vector Synthesis Master library (vs-object_2d-help.pd), available on Github at https://github.com/macumbista. Used with permission.